backend/webapp/generate-contract/requirements.txt
python-docx>=1.1.0
psycopg2-binary>=2.9.0
boto3>=1.26.0
lxml>=4.9.0